update rofi scripts and add new rofi wallpaper selector

+++ private_dot_config/rofi/scripts/executable_powermenu.sh
@@ -1,8 +1,4 @@
-#!/usr/bin/env bash
-
-function rofi_cmd() {
-  inputbar=false rofi -theme menu -dmenu -selected-row 2 -l 5 
-}
+#!/bin/bash
 
 shutdown=""
 reboot=""
@@ -11,22 +7,25 @@ suspend=""
 logout=""
 
 options="$shutdown\n$reboot\n$lock\n$suspend\n$logout"
-chosen="$(echo -e $options | rofi_cmd)"
 
-case "$chosen" in
-  "$shutdown")
-    systemctl poweroff
-    ;;
-  "$reboot")
-    systemctl reboot
-    ;;
-  "$lock")
-    loginctl lock-session
-    ;;
-  "$suspend")
-    systemctl suspend
-    ;;
-  "$logout")
-    swaymsg exit
-    ;;
-esac
+if [[ -z $1 ]]; then
+  echo -en $options
+else
+  case "$1" in
+    "$shutdown")
+      systemctl poweroff
+      ;;
+    "$reboot")
+      systemctl reboot
+      ;;
+    "$lock")
+      loginctl lock-session
+      ;;
+    "$suspend")
+      systemctl suspend
+      ;;
+    "$logout")
+      swaymsg exit
+      ;;
+  esac
+fi
